backend/requirements.txt
flask
flask-cors
crewai
crewai-tools
PyPDF2
python-dotenv
langchain-google-genai
google-generativeai
